Task 6

+++ 0x00-hello_world/6-size.c
@@ -1,18 +1,17 @@
 #include <stdio.h>
 
 /**
- * main - print the string of the function
+ * main - Entry point
  *
- * Description: using the main function
- * this program prints the size of various types on the computer it is compiled and run on
+ * Description: prints the size of various types on the computer it is compiled and run on
  * Return: 0
  */
 int main(void)
 {
-	printf("Size of a char: %d byte(s)\n", (int)sizeof(char));
-	printf("Size of an int: %d byte(s)\n", (int)sizeof(int));
-	printf("Size of a long int: %d byte(s)\n" (int)sizeof(long int));
-	printf("Size of a long long int: %d byte(s)\n" (int)sizeof(long long int));
-	printf("Size of a float: %d byte(s)\n" (int)sizeof(float));
+	printf("Size of a char: %d byte(s)\n", sizeof(char));
+	printf("Size of an int: %d byte(s)\n", sizeof(int));
+	printf("Size of a long int: %d byte(s)\n" sizeof(long int));
+	printf("Size of a long long int: %d byte(s)\n" sizeof(long long int));
+	printf("Size of a float: %d byte(s)\n" sizeof(float));
 	return (0);
 }
